Merge PiotrK branch into master

+++ src/main/java/pl/example/repository/AdminRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.example.models.AdminEntity;
 
 @Repository
-public interface AdminRepository extends CrudRepository<AdminEntity, Long> {
+public interface AdminRepository extends CrudRepository<AdminEntity, Integer> {
 }
