+++ src/main/java/pl/example/repository/SubjectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.example.models.SubjectEntity;
 
 @Repository
-public interface SubjectRepository extends CrudRepository<SubjectEntity, Long> {
+public interface SubjectRepository extends CrudRepository<SubjectEntity, Integer> {
 }
